backend/requirements.txt
flask==2.0.1
flask-cors==3.0.10
flask-sqlalchemy==2.5.1
python-dotenv==0.19.0
requests==2.26.0
websockets==10.0
python-binance==1.0.15
alpha_vantage==2.3.1
psycopg2-binary==2.9.1